+++ lib/iron_goat/lib/iron_goat/lib/src/priv/grid/load_grid.c
